Move upload views to own file
