+++ app/stac_api/urls.py
@@ -12,16 +12,16 @@ from stac_api.views.item import AssetDetail
 from stac_api.views.item import AssetsList
 from stac_api.views.item import ItemDetail
 from stac_api.views.item import ItemsList
-from stac_api.views.views import AssetUploadAbort
-from stac_api.views.views import AssetUploadComplete
-from stac_api.views.views import AssetUploadDetail
-from stac_api.views.views import AssetUploadPartsList
-from stac_api.views.views import AssetUploadsList
-from stac_api.views.views import CollectionAssetUploadAbort
-from stac_api.views.views import CollectionAssetUploadComplete
-from stac_api.views.views import CollectionAssetUploadDetail
-from stac_api.views.views import CollectionAssetUploadPartsList
-from stac_api.views.views import CollectionAssetUploadsList
+from stac_api.views.upload import AssetUploadAbort
+from stac_api.views.upload import AssetUploadComplete
+from stac_api.views.upload import AssetUploadDetail
+from stac_api.views.upload import AssetUploadPartsList
+from stac_api.views.upload import AssetUploadsList
+from stac_api.views.upload import CollectionAssetUploadAbort
+from stac_api.views.upload import CollectionAssetUploadComplete
+from stac_api.views.upload import CollectionAssetUploadDetail
+from stac_api.views.upload import CollectionAssetUploadPartsList
+from stac_api.views.upload import CollectionAssetUploadsList
 from stac_api.views.views import ConformancePageDetail
 from stac_api.views.views import LandingPageDetail
 from stac_api.views.views import SearchList
